==> notebooks/4. Pipeline final/4. Armado de interacciones.feather .ipynb <==
import pandas as pd
pd.set_option("display.max_columns", 500)
pd.set_option("display.max_rows", 500)

table = "TB_INTERACCIONES_1a.tsv"
location = "../../data/raw"
sep = '\t'
encoding = 'latin1'
decimal = ','

usable_cols = ['ID',
    'CIF_ID',
    'FECHA',
    'IN_OUT']

chunks = pd.read_csv(f"{location}/{table}",
                     sep=sep,
                     encoding=encoding,
                     decimal=decimal,
                     chunksize=1000000,
                     iterator=True,
                     usecols = usable_cols)
df = pd.concat(chunks)

c_to_check = ["ID","CIF_ID"]

*Cleaning weird values in ID and CIF_ID* 

df = df[~pd.to_numeric(df['ID'], errors='coerce').isnull()]

to_check = []
for val in df["CIF_ID"].unique():
    try:
        float(val)
    except Exception as e:
        print(e)
        to_check.append(val)

df = df[~df["CIF_ID"].isin(to_check)]

df["CIF_ID"].isna().sum()

df["ID"].isna().sum()

to_check = []
for val in df["ID"].unique():
    try:
        int(val)
    except Exception as e:
        print(e)
        to_check.append(val)

df = df[~df["ID"].isin(to_check)]

df = df.drop(columns='ID').astype({'CIF_ID': 'float64'})

df.reset_index(drop=True).to_feather("interacciones.feather")

df.shape

df['CIF_ID'].nunique()
