==> notebooks/4. Pipeline final/10. Creacion de siniestros por CIF_ID.ipynb <==
import pandas as pd

df_sini = pd.read_feather("siniestros.feather")

df_ids = pd.read_feather("poliza_x_cliente.feather")

Verificación de cruce

len(set(df_ids["NUM_SECU_POL"]).intersection(set(df_sini["NUM_SECU_POL"]))) / len(df_ids)

len(set(df_sini["NUM_SECU_POL"]).intersection(set(df_ids["NUM_SECU_POL"]))) / len(df_sini)

### Agregamos CIF_ID a los Siniestros

df_sini_id = pd.merge(df_sini, df_ids, on=['NUM_SECU_POL'], how='left')

df_sini_id.drop_duplicates(subset=["NUM_SECU_POL", "FEC_DENU_SINI"], keep='last', inplace=True)

### Agrupamos siniestros por periodo

data = {fecha:i + 1 for i,fecha in enumerate(sorted(df_sini_id["FEC_DENU_SINI"].unique()))}

data

df_sini_id["periodo_denu_sini"] = df_sini_id["FEC_DENU_SINI"].map(data)

df_sini_id["periodo_liquidacion_sini"] = df_sini_id["FECHA_LIQUIDACION"].map(data)

df_sini_id["periodo_rechazo_sini"] = df_sini_id["FECHA_RECHAZO"].map(data)

to_pivot = df_sini_id[["CIF_ID","NUM_SECU_POL","periodo_denu_sini","periodo_liquidacion_sini","periodo_rechazo_sini"]]

### Pivots por CIF_ID

pd.set_option("display.max_columns", 500)

df_sini = to_pivot.pivot_table(index='CIF_ID',columns=['periodo_denu_sini'], values=['NUM_SECU_POL','periodo_liquidacion_sini','periodo_rechazo_sini'],aggfunc='count', fill_value=0)

df_sini.head()

Reagrupamos los DF

df_sini = pd.DataFrame(df_sini.to_records())

df_sini.head()

** Cambio de nombre de columnas **

df_sini.columns

df_sini = df_sini.rename(columns=lambda x: x.replace("(","").replace(")","").replace(", ","_").replace("'","").replace("NUM_SECU_POL","periodo_sini"))

df_sini.columns

len(df_sini.columns)

df_sini.to_feather("siniestros_x_cif_id.feather")

df_sini.shape
